+++ Modules/Utils/Sources/PocketCastsUtils/Feature Flags/FeatureFlag.swift
@@ -146,6 +146,12 @@ public enum FeatureFlag: String, CaseIterable {
     /// Use single update query to mark all episodes selected synced
     case markAllSyncedInSingleStatement
 
+    /// Enable the winback screen and flow
+    case winback
+
+    /// Show Manage Downloaded episode banner/modal when running in low space in the device
+    case manageDownloadedEpisodes
+
     /// Uses the episode IDs from the server's response rather than our local database IDs
     case useSyncResponseEpisodeIDs
 
@@ -247,6 +253,10 @@ public enum FeatureFlag: String, CaseIterable {
             true
         case .markAllSyncedInSingleStatement:
             true
+        case .winback:
+            false
+        case .manageDownloadedEpisodes:
+			true
         case .useSyncResponseEpisodeIDs:
             true
         }